requirements.txt
# Core dependencies for EasyDiffusion server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
ruamel.yaml>=0.18.0
torchruntime>=2.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
